requirements.txt
altgraph==0.17.3
asttokens==2.0.8
auto-py-to-exe==2.23.1
backcall==0.2.0
bottle==0.12.23
bottle-websocket==0.2.9
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cssselect2==0.7.0
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
Eel==0.14.0
entrypoints==0.4
executing==1.0.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
html5lib==1.1
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.15.3
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pdfkit==1.0.0
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser==2.21
pydyf==0.3.0
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
pyphen==0.13.0
python-dateutil==2.8.2
pywin32==304
pywin32-ctypes==0.2.0
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
stack-data==0.5.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
Werkzeug==2.2.2
whichcraft==0.6.1
win32printing==0.1.3
yapf==0.32.0
yarg==0.1.9
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.1
